Se termino la implementacion

+++ src/Model/Impl/Iris/IrisPopulation.java
@@ -2,7 +2,7 @@ package Model.Impl.Iris;
 
 import Model.Abstraction.Population;
 
-public class IrisPopulation extends Population<int[]> { // TODO: anzola
+public class IrisPopulation extends Population<double[]> { // TODO: anzola
     public IrisPopulation(int population) throws Exception {
         super(IrisAgent.class, population);
     }
